current ver.

+++ src/db_layer.h
@@ -5,12 +5,12 @@ typedef struct db_token_rec_t {
 	uint64_t cid;
 	uint8_t attr;
 	uint8_t schedule_mask;
-	uint16_t addr;
+	uint32_t addr;
 	uint64_t zone_mask;
 } db_token_rec_t;
 
 typedef struct db_token_addr_t {
-	uint16_t addr;
+	uint32_t addr;
 	uint16_t zone_id;	
 } db_token_addr_t;
 
@@ -21,8 +21,8 @@ int db_sync();
 int db_add_token(db_token_rec_t* rec);
 
 db_token_rec_t* db_get_token_by_c(uint64_t cid);
-db_token_rec_t* db_get_token_by_a(uint16_t addr);
-uint64_t db_get_cid_by_a(uint16_t addr);
+db_token_rec_t* db_get_token_by_a(uint32_t addr);
+uint64_t db_get_cid_by_a(uint32_t addr);
 int db_get_max_addr();
 
 #endif /* _DB_LAYER_H  */
